REF: move bits of IntegerArray/FloatingArray tests to masked_shared
