+++ pandas/tests/extension/masked_shared.py
@@ -1,6 +1,7 @@
 """
 Shared test code for IntegerArray/FloatingArray/BooleanArray.
 """
+import numpy as np
 import pytest
 
 from pandas.compat import (
@@ -14,6 +15,46 @@ from pandas.tests.extension import base
 
 
 class Arithmetic(base.BaseArithmeticOpsTests):
+    def _check_op(self, s, op, other, op_name, exc=NotImplementedError):
+        if exc is None:
+            sdtype = tm.get_dtype(s)
+
+            if hasattr(other, "dtype") and isinstance(other.dtype, np.dtype):
+                if sdtype.kind == "f":
+                    if other.dtype.kind == "f":
+                        # other is np.float64 and would therefore always result
+                        # in upcasting, so keeping other as same numpy_dtype
+                        other = other.astype(sdtype.numpy_dtype)
+
+                else:
+                    # i.e. sdtype.kind in "iu""
+                    if other.dtype.kind in "iu" and sdtype.is_unsigned_integer:
+                        # TODO: comment below is inaccurate; other can be int8
+                        #  int16, ...
+                        #  and the trouble is that e.g. if s is UInt8 and other
+                        #  is int8, then result is UInt16
+                        # other is np.int64 and would therefore always result in
+                        # upcasting, so keeping other as same numpy_dtype
+                        other = other.astype(sdtype.numpy_dtype)
+
+            result = op(s, other)
+            expected = self._combine(s, other, op)
+
+            if sdtype.kind in "iu":
+                if op_name in ("__rtruediv__", "__truediv__", "__div__"):
+                    expected = expected.fillna(np.nan).astype("Float64")
+                else:
+                    # combine method result in 'biggest' (int64) dtype
+                    expected = expected.astype(sdtype)
+            else:
+                # combine method result in 'biggest' (float64) dtype
+                expected = expected.astype(sdtype)
+
+            self.assert_equal(result, expected)
+        else:
+            with pytest.raises(exc):
+                op(s, other)
+
     def check_opname(self, ser: pd.Series, op_name: str, other, exc=None):
         # overwriting to indicate ops don't raise an error
         super().check_opname(ser, op_name, other, exc=None)
