+++ pandas/tests/extension/test_floating.py
@@ -16,11 +16,7 @@ be added to the array-specific tests in `pandas/tests/arrays/`.
 import numpy as np
 import pytest
 
-from pandas.core.dtypes.common import is_extension_array_dtype
-
 import pandas as pd
-import pandas._testing as tm
-from pandas.api.types import is_float_dtype
 from pandas.core.arrays.floating import (
     Float32Dtype,
     Float64Dtype,
@@ -104,28 +100,7 @@ class TestDtype(base.BaseDtypeTests):
 
 
 class TestArithmeticOps(masked_shared.Arithmetic):
-    def _check_op(self, s, op, other, op_name, exc=NotImplementedError):
-        if exc is None:
-            sdtype = tm.get_dtype(s)
-            if (
-                hasattr(other, "dtype")
-                and not is_extension_array_dtype(other.dtype)
-                and is_float_dtype(other.dtype)
-            ):
-                # other is np.float64 and would therefore always result in
-                # upcasting, so keeping other as same numpy_dtype
-                other = other.astype(sdtype.numpy_dtype)
-
-            result = op(s, other)
-            expected = self._combine(s, other, op)
-
-            # combine method result in 'biggest' (float64) dtype
-            expected = expected.astype(sdtype)
-
-            self.assert_equal(result, expected)
-        else:
-            with pytest.raises(exc):
-                op(s, other)
+    pass
 
 
 class TestComparisonOps(masked_shared.Comparison):
